# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-commerce storefront API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface Product {
  id: number;
  slug: string;
  name: string;
  description: string;
  /** @nullable */
  shortDescription?: string | null;
  priceCents: number;
  /** @nullable */
  compareAtCents?: number | null;
  currency: string;
  images: string[];
  inventory: number;
  /** @nullable */
  weightOz?: number | null;
  tags?: string[];
  /** @nullable */
  seoTitle?: string | null;
  /** @nullable */
  seoDescription?: string | null;
  featured: boolean;
  published: boolean;
  createdAt: string;
}

export interface ProductInput {
  slug: string;
  name: string;
  description: string;
  /** @nullable */
  shortDescription?: string | null;
  priceCents: number;
  /** @nullable */
  compareAtCents?: number | null;
  currency?: string;
  images?: string[];
  inventory: number;
  /** @nullable */
  weightOz?: number | null;
  tags?: string[];
  /** @nullable */
  seoTitle?: string | null;
  /** @nullable */
  seoDescription?: string | null;
  featured?: boolean;
  published?: boolean;
}

export interface CartItem {
  id: number;
  productId: number;
  quantity: number;
  product: Product;
  lineTotalCents: number;
}

export interface Cart {
  id: string;
  items: CartItem[];
  subtotalCents: number;
  currency: string;
}

export interface AddressInput {
  name: string;
  street1: string;
  /** @nullable */
  street2?: string | null;
  city: string;
  state: string;
  zip: string;
  country: string;
  /** @nullable */
  phone?: string | null;
}

export interface ShippingRate {
  id: string;
  carrier: string;
  service: string;
  amountCents: number;
  currency: string;
  /** @nullable */
  deliveryDays: number | null;
}

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  /** @nullable */
  productImage?: string | null;
  quantity: number;
  priceCents: number;
}

export interface Order {
  id: number;
  status: string;
  /** @nullable */
  email?: string | null;
  items: OrderItem[];
  subtotalCents: number;
  shippingCents: number;
  taxCents: number;
  totalCents: number;
  currency: string;
  shippingAddress?: AddressInput;
  /** @nullable */
  trackingCode?: string | null;
  /** @nullable */
  labelUrl?: string | null;
  createdAt: string;
}

export interface BlogPost {
  id: number;
  slug: string;
  title: string;
  excerpt: string;
  content: string;
  /** @nullable */
  coverImage?: string | null;
  /** @nullable */
  author?: string | null;
  tags?: string[];
  /** @nullable */
  seoTitle?: string | null;
  /** @nullable */
  seoDescription?: string | null;
  published: boolean;
  /** @nullable */
  publishedAt?: string | null;
  createdAt: string;
}

export interface BlogPostInput {
  slug: string;
  title: string;
  excerpt: string;
  content: string;
  /** @nullable */
  coverImage?: string | null;
  /** @nullable */
  author?: string | null;
  tags?: string[];
  /** @nullable */
  seoTitle?: string | null;
  /** @nullable */
  seoDescription?: string | null;
  published?: boolean;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface AdminStats {
  totalOrders: number;
  ordersThisMonth: number;
  revenueCentsThisMonth: number;
  pendingFulfillment: number;
  lowStock: number;
  totalProducts: number;
  recentOrders?: Order[];
}

/**
 * Not found
 */
export type NotFoundResponse = Error;

export type ListProductsParams = {
  search?: string;
  tag?: string;
  featured?: boolean;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
};

export type GetCartParams = {
  cartId?: string;
};

export type AddCartItemBody = {
  cartId?: string;
  productId: number;
  /** @minimum 1 */
  quantity: number;
};

export type UpdateCartItemBody = {
  cartId: string;
  /** @minimum 0 */
  quantity: number;
};

export type RemoveCartItemParams = {
  cartId: string;
};

export type GetShippingRatesBody = {
  cartId: string;
  address: AddressInput;
};

export type CreateCheckoutBody = {
  cartId: string;
  email?: string;
  address: AddressInput;
  shippingRateId: string;
};

export type CreateCheckout200 = {
  url: string;
  orderId: number;
};

export type ListBlogPostsParams = {
  tag?: string;
  limit?: number;
};

export type ListAdminOrdersParams = {
  status?: string;
};

export type FulfillOrderBody = {
  shippingRateId: string;
};