# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: E-commerce storefront API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /blog/posts:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List published blog posts
      parameters:
        - name: tag
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BlogPost" }
  /blog/posts/{slug}:
    get:
      operationId: getBlogPostBySlug
      tags: [blog]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BlogPost" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /admin/products:
    get:
      operationId: listAdminProducts
      tags: [admin]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
    post:
      operationId: createProduct
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
  /admin/products/{id}:
    patch:
      operationId: updateProduct
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProductInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
    delete:
      operationId: deleteProduct
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }
  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      parameters:
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Order" }
  /admin/orders/{id}/fulfill:
    post:
      operationId: fulfillOrder
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [shippingRateId]
              properties:
                shippingRateId: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
  /admin/blog/posts:
    get:
      operationId: listAdminBlogPosts
      tags: [admin]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BlogPost" }
    post:
      operationId: createBlogPost
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BlogPostInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BlogPost" }
  /admin/blog/posts/{id}:
    patch:
      operationId: updateBlogPost
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BlogPostInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BlogPost" }
    delete:
      operationId: deleteBlogPost
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }
...
components:
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
    Product:
      type: object
      required: [id, slug, name, description, priceCents, currency, images, inventory, featured, published, createdAt]
      properties:
        id: { type: integer }
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        shortDescription: { type: ["string", "null"] }
        priceCents: { type: integer }
        compareAtCents: { type: ["integer", "null"] }
        currency: { type: string }
        images:
          type: array
          items: { type: string }
        inventory: { type: integer }
        weightOz: { type: ["number", "null"] }
        tags:
          type: array
          items: { type: string }
        seoTitle: { type: ["string", "null"] }
        seoDescription: { type: ["string", "null"] }
        featured: { type: boolean }
        published: { type: boolean }
        createdAt: { type: string, format: date-time }
    ProductInput:
      type: object
      required: [slug, name, description, priceCents, inventory]
      properties:
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        shortDescription: { type: ["string", "null"] }
        priceCents: { type: integer }
        compareAtCents: { type: ["integer", "null"] }
        currency: { type: string, default: usd }
        images:
          type: array
          items: { type: string }
        inventory: { type: integer }
        weightOz: { type: ["number", "null"] }
        tags:
          type: array
          items: { type: string }
        seoTitle: { type: ["string", "null"] }
        seoDescription: { type: ["string", "null"] }
        featured: { type: boolean }
        published: { type: boolean }
...
    AddressInput:
      type: object
      required: [name, street1, city, state, zip, country]
      properties:
        name: { type: string }
        street1: { type: string }
        street2: { type: ["string", "null"] }
        city: { type: string }
        state: { type: string }
        zip: { type: string }
        country: { type: string }
        phone: { type: ["string", "null"] }
...
    Order:
      type: object
      required: [id, status, items, subtotalCents, shippingCents, taxCents, totalCents, currency, createdAt]
      properties:
        id: { type: integer }
        status: { type: string }
        email: { type: ["string", "null"] }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        subtotalCents: { type: integer }
        shippingCents: { type: integer }
        taxCents: { type: integer }
        totalCents: { type: integer }
        currency: { type: string }
        shippingAddress: { $ref: "#/components/schemas/AddressInput" }
        trackingCode: { type: ["string", "null"] }
        labelUrl: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    OrderItem:
      type: object
      required: [id, productId, productName, quantity, priceCents]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        productImage: { type: ["string", "null"] }
        quantity: { type: integer }
        priceCents: { type: integer }
    BlogPost:
      type: object
      required: [id, slug, title, excerpt, content, published, createdAt]
      properties:
        id: { type: integer }
        slug: { type: string }
        title: { type: string }
        excerpt: { type: string }
        content: { type: string }
        coverImage: { type: ["string", "null"] }
        author: { type: ["string", "null"] }
        tags:
          type: array
          items: { type: string }
        seoTitle: { type: ["string", "null"] }
        seoDescription: { type: ["string", "null"] }
        published: { type: boolean }
        publishedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    BlogPostInput:
      type: object
      required: [slug, title, excerpt, content]
      properties:
        slug: { type: string }
        title: { type: string }
        excerpt: { type: string }
        content: { type: string }
        coverImage: { type: ["string", "null"] }
        author: { type: ["string", "null"] }
        tags:
          type: array
          items: { type: string }
        seoTitle: { type: ["string", "null"] }
        seoDescription: { type: ["string", "null"] }
        published: { type: boolean }
...