# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-commerce storefront API
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPost {
  id: number;
  slug: string;
  title: string;
  excerpt: string;
  content: string;
  /** @nullable */
  coverImage?: string | null;
  /** @nullable */
  author?: string | null;
  tags?: string[];
  /** @nullable */
  seoTitle?: string | null;
  /** @nullable */
  seoDescription?: string | null;
  published: boolean;
  /** @nullable */
  publishedAt?: Date | null;
  createdAt: Date;
}