# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-commerce storefront API
 * OpenAPI spec version: 0.1.0
 */
import type { AddressInput } from "./addressInput";
import type { OrderItem } from "./orderItem";

export interface Order {
  id: number;
  status: string;
  /** @nullable */
  email?: string | null;
  items: OrderItem[];
  subtotalCents: number;
  shippingCents: number;
  taxCents: number;
  totalCents: number;
  currency: string;
  shippingAddress?: AddressInput;
  /** @nullable */
  trackingCode?: string | null;
  /** @nullable */
  labelUrl?: string | null;
  createdAt: Date;
}