# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-commerce storefront API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  slug: string;
  name: string;
  description: string;
  /** @nullable */
  shortDescription?: string | null;
  priceCents: number;
  /** @nullable */
  compareAtCents?: number | null;
  currency?: string;
  images?: string[];
  inventory: number;
  /** @nullable */
  weightOz?: number | null;
  tags?: string[];
  /** @nullable */
  seoTitle?: string | null;
  /** @nullable */
  seoDescription?: string | null;
  featured?: boolean;
  published?: boolean;
}